Move files as per #243 discussion
