+++ CPP/Clipper2Lib/include/clipper2/clipper.core.h
