+++ CPP/Clipper2Lib/src/clipper.offset.cpp
@@ -9,8 +9,8 @@
 *******************************************************************************/
 
 #include <cmath>
-#include "clipper.h"
-#include "clipper.offset.h"
+#include "clipper2/clipper.h"
+#include "clipper2/clipper.offset.h"
 
 namespace Clipper2Lib {
