+++ CPP/Utils/ClipFileLoad.h
@@ -7,7 +7,7 @@
 
 #include <fstream>
 #include <string>
-#include "clipper.h"
+#include "clipper2/clipper.h"
 #ifndef _WIN32
 #include <sys/types.h>
 #include <sys/stat.h>
